Account for ignored substitutions when checking for malformed tables
Ignored substitutions are internally removed in rstcheck, and this
changes the length of cell contents in tables. This leads to the
detection of malformed tables, which are in fact false positives.

This commit solves the issue by remplacing the substitution reference by
a string of the same length: "|foo|" will be replaced by "xfoox", thus
preserving the length of the padding in the cells.

This patch was inspired by Torbjørn Sørby (@torbsorb).
Fixes: #81